fix: 라운드 횟수 예외처리 추가

+++ src/main/kotlin/racingcar/View/InputView.kt
@@ -21,11 +21,11 @@ class InputView {
         // '.' 줄이기 위해 input함수 따로 작성
         private fun getInput(): String = Console.readLine()
 
-        fun getRoundNumber(): Int {
+        fun getRoundNumber(): Long {
             println(TIME_INPUT_MENTION)
             val number = getInput()
             Exceptions.checkNumber(number)
-            return number.toInt()
+            return number.toLong()
         }
     }
 }
